requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
pydantic==2.9.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
python-jose==3.3.0
python-multipart==0.0.9